+++ src/Pimcore/WithAdminMode.php
@@ -6,6 +6,9 @@ namespace Neusta\Pimcore\TestingFramework\Pimcore;
 
 trait WithAdminMode
 {
+    /** @internal */
+    private static bool $adminModeWasEnabled;
+
     /**
      * @internal
      *
@@ -13,6 +16,19 @@ trait WithAdminMode
      */
     public static function _enableAdminMode(): void
     {
+        self::$adminModeWasEnabled = AdminMode::isEnabled();
         AdminMode::enable();
+    }
+
+    /**
+     * @internal
+     *
+     * @afterClass
+     */
+    public static function _resetAdminMode(): void
+    {
+        if (false === self::$adminModeWasEnabled) {
+            AdminMode::disable();
+        }
     }
 }