+++ src/Pimcore/WithoutAdminMode.php
@@ -6,6 +6,9 @@ namespace Neusta\Pimcore\TestingFramework\Pimcore;
 
 trait WithoutAdminMode
 {
+    /** @internal */
+    private static bool $adminModeWasEnabled;
+
     /**
      * @internal
      *
@@ -13,6 +16,19 @@ trait WithoutAdminMode
      */
     public static function _disableAdminMode(): void
     {
+        self::$adminModeWasEnabled = AdminMode::isEnabled();
         AdminMode::disable();
+    }
+
+    /**
+     * @internal
+     *
+     * @afterClass
+     */
+    public static function _resetAdminMode(): void
+    {
+        if (true === self::$adminModeWasEnabled) {
+            AdminMode::enable();
+        }
     }
 }